+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.0 (DEV)
+
+- Fix: Some text grammar.
+- Add: Display hovered cell coordinates in footer.
+- Add: Open first level when a file opened or created.
+
 ## 0.1.0 (2022-10-13)
 
 - MVP is reached. First public release.